Fix pagination logic
Turns out :has_more was not the best option to test if pagination was necessary.

:has_more works for time based pagination, but since we are using cursors (the
recommended way), it is best to test the presence of a cursor in the response.

If it is nil it means we are not even paginating (there are only a small number
of results) and if it is an empty string, it means we have reached the end of
the pagination.