Improve check for cursor presence using nil punning
I think having both tests makes the code a bit more intelligible because there
are two scenarios here. In the first, the cursor doesn't exist at all, and it
means we are not even paginating and should just return what we've got in our
very first answer. In the second one, the cursor exists but points to an empty
_string_, which means we have exhausted the cursors (finished pagination) and
the last answer contained the last items. In this case we should return the
accumulated value retrieved from all the previous answers.

With that said, knowing that `empty?` returns true for both cases just
simplifies the code and abstracts the implementation details of the Slack API:
we have reached the place where we return what we have. And it is just more
Lispy and even adds cool terminology props. :3 Thanks Arne!